% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bootstrap.R
\name{add_bootstrap}
\alias{add_bootstrap}
\title{An extension of nodelabels to add symbols indicating the bootstrap support pass certain threshold}
\usage{
add_bootstrap(phy, pch, threshold)
}
\arguments{
\item{phy}{A tree object with bootstrap support values in nodes.}

\item{pch}{Shape to add}

\item{threshold}{A threshold value. Only the nodes that have support higher than this number will be plotted.}
}
\value{
plots shapes indicating bootstrap support in the current plot window
}
\description{
An extension of nodelabels to add symbols indicating the bootstrap support pass certain threshold
}
