% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_clades.R
\name{annotate_clades}
\alias{annotate_clades}
\title{Plot and label clades of a tree}
\usage{
annotate_clades(phy, annot_table)
}
\arguments{
\item{phy}{A tree object with bootstrap support values as node labels.}

\item{annot_table}{A data frame that has each taxa in monophyletic clades assigned to a group. The first column name must be taxa name and the second column must be clade names and name clade.name.}
}
\value{
A plot object
}
\description{
Plot and label clades of a tree
}
