% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_table.R
\name{annotate_table}
\alias{annotate_table}
\title{Assigns unassigned taxa in a clade to a group based on assignment of other taxa from the table}
\usage{
annotate_table(phy, annot_table)
}
\arguments{
\item{phy}{A rooted tree object.}

\item{annot_table}{A data frame where each row has a taxa and corresponding groups. Groups must be in column labeled `Phylo.group` and the first column must be taxa name in `phy`.}
}
\value{
A dataframe where all the taxa are now assigned to a group based on monophylyl.
}
\description{
Assigns unassigned taxa in a clade to a group based on assignment of other taxa from the table
}
