% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backbone_tree.R
\name{backbone_tree}
\alias{backbone_tree}
\title{Drops, all but one taxa from a monophyletic clade, keeps the info on total number of taxa}
\usage{
backbone_tree(df, phy, size = TRUE, backbone = TRUE)
}
\arguments{
\item{df}{A dataframe with two columns, first column has taxa name matching phy, and second column has its group affiliation.}

\item{phy}{A tree object.}

\item{size}{If set to TRUE draws clades proportional to number of taxa. If FALSE, draws clades that are of same size. (default=TRUE)}

\item{backbone}{If set to TRUE return backbone tree, else returns a regular tree with dropped branches. (default=TRUE)}
}
\value{
A backbone tree with
}
\description{
Drops, all but one taxa from a monophyletic clade, keeps the info on total number of taxa
}
