% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makebackbone.R
\name{makebackbone}
\alias{makebackbone}
\title{Collapses monophyletic clade if they belong to a group}
\usage{
makebackbone(df, phy, plot_type)
}
\arguments{
\item{df}{A dataframe with two columns, first column has taxa name matching phy, and second column has its group affiliation.}

\item{phy}{A tree object.}

\item{plot_type}{Indicate whether the collapsed branches' triangle size should represent the number of taxa (size) or not (same).}
}
\value{
backbone phylogeny object
}
\description{
Collapses monophyletic clade if they belong to a group
}
